requirements.txt
fastapi
uvicorn[standard]
sqlmodel>=0.0.14
pydantic[email]>=2.0
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
psycopg2-binary
sqlalchemy>=2.0
